requirements.txt
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
gensim==4.3.3
imbalanced-learn==0.13.0
imblearn==0.0
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.0
nltk==3.9.1
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
regex==2024.11.6
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
smart-open==7.1.0
threadpoolctl==3.5.0
tqdm==4.67.1
tzdata==2025.1
wordcloud==1.9.4
wrapt==1.17.2
